requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.21.18
botocore==1.24.18
cached-property==1.5.2
celery==5.1.2
celery-progress==0.1.2
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextvars==2.4
decorator==4.4.2
deep-translator==1.5.0
Deprecated==1.2.13
Django==3.2.10
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.13.1
google-trans-new==1.1.9
googletrans==3.0.0
gTTS==2.2.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.13.3
imageio-ffmpeg==0.4.5
immutables==0.16
importlib-metadata==4.8.3
jmespath==0.10.0
kombu==5.1.0
libretranslatepy==2.1.1
lxml==4.7.1
moviepy==1.0.3
numpy==1.19.5
packaging==21.3
Pillow==8.4.0
proglog==0.1.9
prompt-toolkit==3.0.26
psycopg2==2.7.7
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
pyttsx3 @ git+https://github.com/nateshmbhat/pyttsx3@5d3755b060a980f48fcaf81df018dd06cbd17a8f
pytz==2021.3
redis==4.1.2
requests==2.26.0
rfc3986==1.5.0
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
translate==3.6.1
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
